feat: Phase 2 - queue, persistence, i18n, user prefs, stats, rate-limit, default quality UX
- Redis service with graceful fallback (url cache, rate-limit, arq enqueue, stats)
- User/Job services with DB persistence (SQLite/Postgres) and fixed Job autoincrement
- UserMiddleware + I18nMiddleware (DB-backed lang) + Redis-backed rate limiter
- Keyboards: Redis cache, default-quality highlight, language selector
- Handlers: url_handler caches async+DB upsert+default quality, callback enqueues or direct fallback with Job tracking
- Start/settings/stats: real DB stats, language+quality persistence
- Workers: Job status tracking, increment counters, progress throttle
- Main: healthchecks (/health,/healthz), cleanup lifecycle, USE_QUEUE flag
- Locales: es/fr/ar + en/hi updates, docker-compose USE_QUEUE=true, .env.example docs
- Tests: 11 new Phase-2 tests (17 total passing)

+++ app/models/job.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 from datetime import datetime
-from sqlalchemy import BigInteger, String, DateTime, func, Text
+
+from sqlalchemy import BigInteger, DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.services.db import Base
@@ -10,7 +11,8 @@ from app.services.db import Base
 class Job(Base):
     __tablename__ = "jobs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    # Use Integer for SQLite autoincrement compatibility (works on Postgres too for <2B rows)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(BigInteger)
     chat_id: Mapped[int] = mapped_column(BigInteger)
     url: Mapped[str] = mapped_column(Text)
